+++ src/javascript/customization.js
@@ -26,7 +26,6 @@ const loadTheme = number => {
   if(themes[i].type=="custom_theme"){
     plugins.enableCSS(themes[i])
     themeObject = themes[i]
-    document.documentElement.removeAttribute("style");
   }
   themeObject = themes[i];
   const colors = themes[i]["colors"]; //Take the colors object inside the json file of the selected theme
@@ -39,7 +38,7 @@ const loadTheme = number => {
         i+=2;
       } catch {}
     } else {
-      if ((current_config.animationsPreferences == "desactivated" && Object.keys(colors)[i] != "scalation") || current_config.animationsPreferences == "activated") { //Prevent changing the scalation when the animations are off
+      if ((current_config.animationsPreferences == "desactivated" && Object.keys(colors)[i] != "scalation") ||  current_config.animationsPreferences == "activated" || Object.keys(colors)[i] != "blur" ) { //Prevent changing the scalation when the animations are off
         document.documentElement.style.setProperty("--" + Object.keys(colors)[i], colors[Object.keys(colors)[i]]); //Update the CSS variables
       }
     }
@@ -47,7 +46,6 @@ const loadTheme = number => {
   for (i = 0; i < editors.length; i++) {
     if (editors[i].editor != undefined) editors[i].editor.setOption("theme", themes[number]["highlight"]); //Update highlither after applying a new theme
   }
-  if(themes[i].type=="custom_theme"){
     for(i=0;i<editor_screens.length;i++){
       if (editor_screens[i] != undefined) {
         if (editor_screens[i].terminal != undefined) {
@@ -57,7 +55,6 @@ const loadTheme = number => {
           })
         }
       }
-    }
   }
   current_config.theme = themes[number];
   saveConfig(); //Save the current configuration
@@ -86,7 +83,7 @@ const setThemeByName = name => {
             i+=2;
           } catch {}
         } else {
-          if ((current_config.animationsPreferences == "desactivated" && Object.keys(colors)[i] != "scalation") || current_config.animationsPreferences == "activated") { //Prevent changing the scalation when the animations are off
+          if ((current_config.animationsPreferences == "desactivated" && Object.keys(colors)[i] != "scalation") || current_config.animationsPreferences == "activated" || Object.keys(colors)[i] != "blur" ) { //Prevent changing the scalation when the animations are off
             document.documentElement.style.setProperty("--" + Object.keys(colors)[i], colors[Object.keys(colors)[i]]); //Update the CSS variables
           }
         }