+++ src/javascript/market.js
@@ -226,7 +226,9 @@ const extensions ={
         }else{
           const request = require("request");
           request(`https://raw.githubusercontent.com/${plugin.repo.git.owner.login}/${plugin.repo.git.name}/${plugin.repo.git.default_branch}/readme.md`, function (error, response, body3) {
-            document.getElementById(data.getAttribute('name')+'_div').innerHTML += `<div class=ext_content>${!error?marked(body3):getTranslation("NoReadme")}</div>`
+            console.log(error);
+            console.log(response);
+            document.getElementById(data.getAttribute('name')+'_div').innerHTML += `<div class=ext_content>${!error && response.statusCode ==200?marked(body3):getTranslation("NoReadme")}</div>`
           })
         }
     },
@@ -254,6 +256,8 @@ const extensions ={
     },
     updateExtension: function(name){
       const plugin = graviton.getPlugin(name)
+      const nodegit = require("nodegit");
+      /*
       const new_update = plugin.local!=undefined?getVersionSum(plugin.repo.package.version)>getVersionSum(plugin.local.version):false;
       if (!fs.existsSync(path.join(plugins_folder,name))) {
         new Notification('Market',name+ current_config.language["ExtNotInstalled"]);
@@ -280,8 +284,24 @@ const extensions ={
         }else{
           plugins.install(plugin.repo.package)
         }
-      });
-      
+      });*/
+      const repositoryPath = path.join(plugins_folder.replace(/\\/g, '\\\\'),plugin.local.name);
+      const remoteName = 'origin';
+      const branch = 'master';
+      pull(repositoryPath,remoteName,branch,function(a,b){
+        if(a!=null)return;
+        console.log(a,b);
+        const updated_ext_event = new CustomEvent("updated_installed",{
+          detail:{
+            name : name
+          }
+        })
+        document.dispatchEvent(updated_ext_event);
+        new Notification('Market',name+ current_config.language["ExtUpdated"]);
+        if(plugin.repo.package["dependencies"]!=undefined){
+          plugins.installDependencies(plugin.repo.package);
+        }
+      }) 
     },
     uninstallExtension: function(name){
       const rimraf = require('rimraf');
@@ -384,6 +404,7 @@ const extensions ={
       })
     },
     disableCSS: function(config){
+      if(config.css==undefined ||config.css.length == 0) return;
       const csss = document.getElementsByClassName(config.name+"_css");
       for(b=0;b<csss.length;b++){
         csss[b].remove();
@@ -391,6 +412,7 @@ const extensions ={
       }   
     },
     enableCSS: function(config){
+      if(config.css==undefined ||config.css.length == 0) return;
       for (b = 0; b < config.css.length; b++) {
         const link = document.createElement("link");
         link.setAttribute("rel", "stylesheet");
@@ -411,3 +433,34 @@ const extensions ={
       return parseInt(result,10);
    })(input)
  }
+
+const installCli = function(){
+  const npm = require('npm')
+  npm.load({
+    global:true
+  },function (er) {
+    if (er) return er;
+    npm.commands.install(['graviton-cli'], function (er, data) {
+      if (er) return er;
+      console.log("cli installed!")
+    })
+  })
+}
+
+function pull(path, remote, branch, call) {
+    const Git = require('nodegit');
+    const open = Git.Repository.open;   
+    let repository;
+    const remoteBranch = remote + '/' + branch;
+    open(path)
+      .then(function (_repository) {
+          repository = _repository;
+          return repository.fetch(remote);
+      }, call)
+      .then(function () {
+          return repository.mergeBranches(branch, remoteBranch);
+      }, call)
+      .then(function (oid) {
+        call(null, oid);
+      }, call);
+};